requirements.txt
numpy
scipy
lxml
matplotlib
opencv-python
tqdm
dotmap
GPUtil
psutil
colored
torch
tabulate
open3d
transforms3d
easydict
visdom
ray
GPUtil
protobuf
wandb
ray
drake
gym
hydra
hydra-core==1.3
